Introduce personality modules.
Backport of 0.19 personality modules without the xml config file

The attempt to merge xml config file as well as personality modules
into 0.19 branch was just too great a change.

Keeping the 0.19 branch for the xml config changes.

+++ usr/smc.h
@@ -25,13 +25,13 @@
 #define START_MAP	0x0200
 #define START_STORAGE	0x0400
 
-int smc_allow_removal(struct scsi_cmd *cmd);
-int smc_initialize_element(struct scsi_cmd *cmd);
-int smc_initialize_element_range(struct scsi_cmd *cmd);
-int smc_move_medium(struct scsi_cmd *cmd);
-int smc_read_element_status(struct scsi_cmd *cmd);
-int smc_rezero(struct scsi_cmd *cmd);
-int smc_start_stop(struct scsi_cmd *cmd);
+uint8_t smc_allow_removal(struct scsi_cmd *cmd);
+uint8_t smc_initialize_element(struct scsi_cmd *cmd);
+uint8_t smc_initialize_element_range(struct scsi_cmd *cmd);
+uint8_t smc_move_medium(struct scsi_cmd *cmd);
+uint8_t smc_read_element_status(struct scsi_cmd *cmd);
+uint8_t smc_rezero(struct scsi_cmd *cmd);
+uint8_t smc_start_stop(struct scsi_cmd *cmd);
 
 int slotOccupied(struct s_info *s);
 void setImpExpStatus(struct s_info *s, int flg);
