+++ usr/spc.h
@@ -9,20 +9,20 @@ extern uint32_t SPR_Reservation_Generation;
 extern uint8_t SPR_Reservation_Type;
 
 
-int resp_spc_pro(uint8_t *cdb, struct vtl_ds *dbuf_p);
-int resp_spc_pri(uint8_t *cdb, struct vtl_ds *dbuf_p);
+uint8_t resp_spc_pro(uint8_t *cdb, struct vtl_ds *dbuf_p);
+uint8_t resp_spc_pri(uint8_t *cdb, struct vtl_ds *dbuf_p);
 
-int spc_illegal_op(struct scsi_cmd *cmd);
-int spc_inquiry(struct scsi_cmd *cmd);
-int spc_log_select(struct scsi_cmd *cmd);
-int spc_log_sense(struct scsi_cmd *cmd);
-int spc_mode_select(struct scsi_cmd *cmd);
-int spc_mode_sense(struct scsi_cmd *cmd);
-int spc_recv_diagnostics(struct scsi_cmd *cmd);
-int spc_release(struct scsi_cmd *cmd);
-int spc_request_sense(struct scsi_cmd *cmd);
-int spc_reserve(struct scsi_cmd *cmd);
-int spc_send_diagnostics(struct scsi_cmd *cmd);
-int spc_tur(struct scsi_cmd *cmd);
+uint8_t spc_illegal_op(struct scsi_cmd *cmd);
+uint8_t spc_inquiry(struct scsi_cmd *cmd);
+uint8_t spc_log_select(struct scsi_cmd *cmd);
+uint8_t spc_log_sense(struct scsi_cmd *cmd);
+uint8_t spc_mode_select(struct scsi_cmd *cmd);
+uint8_t spc_mode_sense(struct scsi_cmd *cmd);
+uint8_t spc_recv_diagnostics(struct scsi_cmd *cmd);
+uint8_t spc_release(struct scsi_cmd *cmd);
+uint8_t spc_request_sense(struct scsi_cmd *cmd);
+uint8_t spc_reserve(struct scsi_cmd *cmd);
+uint8_t spc_send_diagnostics(struct scsi_cmd *cmd);
+uint8_t spc_tur(struct scsi_cmd *cmd);
 
 #endif	/* SPC_H */
